Add state-space to the code
I added the option to disable state-space in case it won't work, so we still have stable code for the season (always be prepared😂). Also, we never used state-space in the flywheel, so until it stable we need to be sure. We Use kalman and LQR because wpilib force us to use them, and most likely it won't harm