Create setVelocity and stop functions in Flywheel subsystem
The set velocity function has 2 versions, one for state-space and the other is the "native" way (just PID). At the beginning of the state-space part we tell the state-space to update based on the velocity, afterwards we correct the model based on the velocity, and predict the next velocity we should give to the motor. Finally we power the motors