Remove useStateSpace boolean from Shooter class and split the modules
We have already tested the state-space and is better than PID, so we chose to use it as the default in the shooter. Also, the modules are specific to our robot, so I chose to create an object for each module instead of using array. More then that, most likely that no one will create a shooter with more than 2 wheels in FRC.